Port BasicEndConstruct to the new framework
